move all 'completion' code to file XKCD.swift
We move all the 'completion' code to the renamed file 'xkcd.async.swift'
-> 'XKCD.swift'.

We want to keep all the !xkcd-command related code in one file and
change the filename fo follow the filename convention.
